Scripts/Icecrown Citadel: Compile warning fixes

+++ src/server/scripts/Northrend/IcecrownCitadel/boss_rotface.cpp
@@ -233,7 +233,7 @@ class boss_rotface : public CreatureScript
 
         private:
             uint32 infectionCooldown;
-            uint8 infectionStage;
+            uint32 infectionStage;
         };
 
         CreatureAI* GetAI(Creature* creature) const
